Completed the task to find overlapping events within our custom Nano AOD and the centralized created NANOAOD.
Due to different compressions between our and the central NANOAODs we need to find out which overlap both have and filter them out.
This task finds the events that overlap and savin the unique event identifier as tuple in a json.
This json also contains the relative number of overlap as information.